third_party/WebKit/LayoutTests/inspector/help/release-note-expected.txt

Test release note

Browser version: 99

Dumping release note text:
Chrome has been updated to version 99.
Here's what's new in DevTools:
Improved Performance and Memory panels
Edit cookies directly from the Application panel

Learn moreClose
Last version of release note seen should be updated:
99

Click on hide button
Hiding view: release-note
